Fix latitude and longitude tests after moving to geo provider
