+++ faker/providers/address/de_AT/__init__.py
@@ -156,9 +156,3 @@ class Provider(AddressProvider):
 
     def city_with_postcode(self):
         return self.postcode() + " " + self.random_element(self.cities)
-
-    def latitude(self):
-        return self.geo_coordinate(center=47.60707, radius=1)
-
-    def longitude(self):
-        return self.geo_coordinate(center=13.37208, radius=2)
